requirements.txt
absl-py==2.3.0
accelerate==1.9.0
annotated-types==0.7.0
anywidget==0.9.18
asttokens==3.0.0
bbox_visualizer==0.2.2
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
executing==2.2.0
fairscale==0.4.13
filelock==3.18.0
filetype==1.2.0
filterpy==1.4.5
fonttools==4.58.4
fsspec==2025.5.1
ftfy==6.3.1
gdown==5.2.0
grpcio==1.73.0
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.7
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_bbox_widget==0.6.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.1
open_clip_torch==2.32.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.0
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.2.1
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
polygraphy==0.49.26
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psygnal==0.14.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylabel==0.1.55
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rf100vl==1.0.0
rfdetr==1.2.1
roboflow==1.2.2
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
supervision==0.26.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchreid==0.2.5
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
